resources/test/tests/add_cleanup_sync_queue.html

<!DOCTYPE HTML>
<html>
<head>
<title>Test#add_cleanup: queuing synchronous tests</title>
<script src="../../testharness.js"></script>
<script src="../../testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>

<script>
"use strict";
var firstCleaned = false;

test(function(t) {
  test(function() {
    assert_true(
      firstCleaned, "should not execute until first test is complete"
    );
  }, "test defined when no tests are queued, but one test is executing");

  t.add_cleanup(function() {
    firstCleaned = true;
  });
}, "Test with a 'cleanup' function");
</script>
<script type="text/json" id="expected">
{
  "summarized_status": {
    "message": null,
    "stack": null,
    "status_string": "OK"
  },
  "summarized_tests": [
    {
      "stack": null,
      "message": null,
      "name": "Test with a 'cleanup' function",
      "status_string": "PASS",
      "properties": {}
    },
    {
      "stack": null,
      "message": null,
      "name": "test defined when no tests are queued, but one test is executing",
      "status_string": "PASS",
      "properties": {}
    }
  ],
  "type": "complete"
}
</script>
</body>
</html>
